feature: submit action and factory for code runner
Now code runner works with a factory depending the engine required by the client side
New endpoint to submit challenge success

+++ database/migrations/2014_02_03_095849_create_genders_table.php
@@ -13,10 +13,13 @@ class CreateGendersTable extends Migration
      */
     public function up()
     {
-        Schema::create('genders', function (Blueprint $table) {
-            $table->id();
-            $table->string('name', 50);
-        });
+        if (env('APP_ENV') == 'testing' || env('APP_ENV') == 'local') {
+
+            Schema::create('genders', function (Blueprint $table) {
+                $table->id();
+                $table->string('name', 50);
+            });
+        }
     }
 
     /**
@@ -26,6 +29,9 @@ class CreateGendersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('genders');
+        if (env('APP_ENV') == 'testing' || env('APP_ENV') == 'local') {
+
+            Schema::dropIfExists('genders');
+        }
     }
 }
